+++ tests/test_triggers.py
@@ -11,7 +11,8 @@ from apscheduler.triggers.interval import IntervalTrigger
 class TestCronTrigger(object):
     def test_cron_trigger_1(self, timezone):
         trigger = CronTrigger(year='2009/2', month='1/3', day='5-13', timezone=timezone)
-        assert repr(trigger) == "<CronTrigger (year='2009/2', month='1/3', day='5-13')>"
+        assert repr(trigger) == ("<CronTrigger (year='2009/2', month='1/3', day='5-13', "
+                                 "timezone='Europe/Berlin')>")
         assert str(trigger) == "cron[year='2009/2', month='1/3', day='5-13']"
         start_date = timezone.localize(datetime(2008, 12, 1))
         correct_next_date = timezone.localize(datetime(2009, 1, 5))
@@ -25,25 +26,28 @@ class TestCronTrigger(object):
 
     def test_cron_trigger_3(self, timezone):
         trigger = CronTrigger(year='2009', month='2', hour='8-10', timezone=timezone)
-        assert repr(trigger) == "<CronTrigger (year='2009', month='2', hour='8-10')>"
+        assert repr(trigger) == ("<CronTrigger (year='2009', month='2', hour='8-10', "
+                                 "timezone='Europe/Berlin')>")
         start_date = timezone.localize(datetime(2009, 1, 1))
         correct_next_date = timezone.localize(datetime(2009, 2, 1, 8))
         assert trigger.get_next_fire_time(None, start_date) == correct_next_date
 
     def test_cron_trigger_4(self, timezone):
         trigger = CronTrigger(year='2012', month='2', day='last', timezone=timezone)
-        assert repr(trigger) == "<CronTrigger (year='2012', month='2', day='last')>"
+        assert repr(trigger) == ("<CronTrigger (year='2012', month='2', day='last', "
+                                 "timezone='Europe/Berlin')>")
         start_date = timezone.localize(datetime(2012, 2, 1))
         correct_next_date = timezone.localize(datetime(2012, 2, 29))
         assert trigger.get_next_fire_time(None, start_date) == correct_next_date
 
     def test_cron_zero_value(self, timezone):
         trigger = CronTrigger(year=2009, month=2, hour=0, timezone=timezone)
-        assert repr(trigger) == "<CronTrigger (year='2009', month='2', hour='0')>"
+        assert repr(trigger) == ("<CronTrigger (year='2009', month='2', hour='0', "
+                                 "timezone='Europe/Berlin')>")
 
     def test_cron_year_list(self, timezone):
         trigger = CronTrigger(year='2009,2008', timezone=timezone)
-        assert repr(trigger) == "<CronTrigger (year='2009,2008')>"
+        assert repr(trigger) == "<CronTrigger (year='2009,2008', timezone='Europe/Berlin')>"
         assert str(trigger) == "cron[year='2009,2008']"
         start_date = timezone.localize(datetime(2009, 1, 1))
         correct_next_date = timezone.localize(datetime(2009, 1, 1))
@@ -53,7 +57,8 @@ class TestCronTrigger(object):
         trigger = CronTrigger(year='2009', month='2', hour='8-10',
                               start_date='2009-02-03 11:00:00', timezone=timezone)
         assert repr(trigger) == ("<CronTrigger (year='2009', month='2', hour='8-10', "
-                                 "start_date='2009-02-03 11:00:00 CET')>")
+                                 "start_date='2009-02-03 11:00:00 CET', "
+                                 "timezone='Europe/Berlin')>")
         assert str(trigger) == "cron[year='2009', month='2', hour='8-10']"
         start_date = timezone.localize(datetime(2009, 1, 1))
         correct_next_date = timezone.localize(datetime(2009, 2, 4, 8))
@@ -77,7 +82,7 @@ class TestCronTrigger(object):
     def test_cron_weekday_overlap(self, timezone):
         trigger = CronTrigger(year=2009, month=1, day='6-10', day_of_week='2-4', timezone=timezone)
         assert repr(trigger) == ("<CronTrigger (year='2009', month='1', day='6-10', "
-                                 "day_of_week='2-4')>")
+                                 "day_of_week='2-4', timezone='Europe/Berlin')>")
         assert str(trigger) == "cron[year='2009', month='1', day='6-10', day_of_week='2-4']"
         start_date = timezone.localize(datetime(2009, 1, 1))
         correct_next_date = timezone.localize(datetime(2009, 1, 7))
@@ -86,7 +91,7 @@ class TestCronTrigger(object):
     def test_cron_weekday_nomatch(self, timezone):
         trigger = CronTrigger(year=2009, month=1, day='6-10', day_of_week='0,6', timezone=timezone)
         assert repr(trigger) == ("<CronTrigger (year='2009', month='1', day='6-10', "
-                                 "day_of_week='0,6')>")
+                                 "day_of_week='0,6', timezone='Europe/Berlin')>")
         assert str(trigger) == "cron[year='2009', month='1', day='6-10', day_of_week='0,6']"
         start_date = timezone.localize(datetime(2009, 1, 1))
         correct_next_date = None
@@ -94,7 +99,8 @@ class TestCronTrigger(object):
 
     def test_cron_weekday_positional(self, timezone):
         trigger = CronTrigger(year=2009, month=1, day='4th wed', timezone=timezone)
-        assert repr(trigger) == "<CronTrigger (year='2009', month='1', day='4th wed')>"
+        assert repr(trigger) == ("<CronTrigger (year='2009', month='1', day='4th wed', "
+                                 "timezone='Europe/Berlin')>")
         assert str(trigger) == "cron[year='2009', month='1', day='4th wed']"
         start_date = timezone.localize(datetime(2009, 1, 1))
         correct_next_date = timezone.localize(datetime(2009, 1, 28))
@@ -102,7 +108,8 @@ class TestCronTrigger(object):
 
     def test_week_1(self, timezone):
         trigger = CronTrigger(year=2009, month=2, week=8, timezone=timezone)
-        assert repr(trigger) == "<CronTrigger (year='2009', month='2', week='8')>"
+        assert repr(trigger) == ("<CronTrigger (year='2009', month='2', week='8', "
+                                 "timezone='Europe/Berlin')>")
         assert str(trigger) == "cron[year='2009', month='2', week='8']"
         start_date = timezone.localize(datetime(2009, 1, 1))
         correct_next_date = timezone.localize(datetime(2009, 2, 16))
@@ -110,7 +117,8 @@ class TestCronTrigger(object):
 
     def test_week_2(self, timezone):
         trigger = CronTrigger(year=2009, week=15, day_of_week=2, timezone=timezone)
-        assert repr(trigger) == "<CronTrigger (year='2009', week='15', day_of_week='2')>"
+        assert repr(trigger) == ("<CronTrigger (year='2009', week='15', day_of_week='2', "
+                                 "timezone='Europe/Berlin')>")
         assert str(trigger) == "cron[year='2009', week='15', day_of_week='2']"
         start_date = timezone.localize(datetime(2009, 1, 1))
         correct_next_date = timezone.localize(datetime(2009, 4, 8))
@@ -119,7 +127,7 @@ class TestCronTrigger(object):
     def test_cron_extra_coverage(self, timezone):
         # This test has no value other than patching holes in test coverage
         trigger = CronTrigger(day='6,8', timezone=timezone)
-        assert repr(trigger) == "<CronTrigger (day='6,8')>"
+        assert repr(trigger) == "<CronTrigger (day='6,8', timezone='Europe/Berlin')>"
         assert str(trigger) == "cron[day='6,8']"
         start_date = timezone.localize(datetime(2009, 12, 31))
         correct_next_date = timezone.localize(datetime(2010, 1, 6))
@@ -135,7 +143,7 @@ class TestCronTrigger(object):
 
         """
         trigger = CronTrigger(hour='5-6', timezone=timezone)
-        assert repr(trigger) == "<CronTrigger (hour='5-6')>"
+        assert repr(trigger) == "<CronTrigger (hour='5-6', timezone='Europe/Berlin')>"
         assert str(trigger) == "cron[hour='5-6']"
         start_date = timezone.localize(datetime(2009, 9, 25, 7))
         correct_next_date = timezone.localize(datetime(2009, 9, 26, 5))
@@ -166,7 +174,8 @@ class TestCronTrigger(object):
     def test_different_tz(self, timezone):
         alter_tz = pytz.FixedOffset(-600)
         trigger = CronTrigger(year=2009, week=15, day_of_week=2, timezone=timezone)
-        assert repr(trigger) == "<CronTrigger (year='2009', week='15', day_of_week='2')>"
+        assert repr(trigger) == ("<CronTrigger (year='2009', week='15', day_of_week='2', "
+                                 "timezone='Europe/Berlin')>")
         assert str(trigger) == "cron[year='2009', week='15', day_of_week='2']"
         start_date = alter_tz.localize(datetime(2008, 12, 31, 22))
         correct_next_date = timezone.localize(datetime(2009, 4, 8))
@@ -310,7 +319,8 @@ class TestIntervalTrigger(object):
 
     def test_repr(self, trigger):
         assert repr(trigger) == ("<IntervalTrigger (interval=datetime.timedelta(0, 1), "
-                                 "start_date='2009-08-04 00:00:02 CEST')>")
+                                 "start_date='2009-08-04 00:00:02 CEST', "
+                                 "timezone='Europe/Berlin')>")
 
     def test_str(self, trigger):
         assert str(trigger) == "interval[0:00:01]"